+++ appinfo/routes.php
@@ -12,5 +12,7 @@ return [
 		['name' => 'RulesApi#index', 'url' => '/api/v1/rules', 'verb' => 'GET'],
 		['name' => 'RulesApi#update', 'url' => '/api/v1/rules', 'verb' => 'PUT'],
 		['name' => 'RulesApi#simulate', 'url' => '/api/v1/simulate', 'verb' => 'POST'],
+		['name' => 'RulesApi#targets', 'url' => '/api/v1/targets', 'verb' => 'GET'],
+		['name' => 'RulesApi#providers', 'url' => '/api/v1/providers', 'verb' => 'GET'],
 	],
 ];